Library Reference

Elimination Algorithms

CliqueTrees.EliminationAlgorithm — Type

EliminationAlgorithm

A graph elimination algorithm computes a permutation of the vertices of a graph, which induces a chordal completion of the graph. The algorithms below generally seek to minimize the fill (number of edges) or width (largest clique) of the completed graph.

type	name	time	space

BFS	breadth-first search	O(m + n)	O(n)

MCS	maximum cardinality search	O(m + n)	O(n)

LexBFS	lexicographic breadth-first search	O(m + n)	O(m + n)

RCMMD	reverse Cuthill-Mckee (minimum degree)	O(m + n)	O(m + n)

RCMGL	reverse Cuthill-Mckee (George-Liu)	O(m + n)	O(m + n)

MCSM	maximum cardinality search (minimal)	O(mn)	O(n)

LexM	lexicographic breadth-first search (minimal)	O(mn)	O(n)

AMF	approximate minimum fill	O(mn)	O(m + n)

MF	minimum fill	O(mn²)

MMD	multiple minimum degree	O(mn²)	O(m + n)

MinimalChordal	MinimalChordal

CompositeRotations	elimination tree rotation	O(m + n)	O(m + n)

RuleReduction	treewith-safe rule-based reduction

ComponentReduction	connected component reduction

The following additional algorithms are implemented as package extensions and require loading an additional package.

type	name	time	space	package

AMD	approximate minimum degree	O(mn)	O(m + n)	AMD.jl

SymAMD	column approximate minimum degree	O(mn)	O(m + n)	AMD.jl

METIS	multilevel nested dissection			Metis.jl

Spectral	spectral ordering			Laplacians.jl

BT	Bouchitte-Todinca			TreeWidthSolver.jl

Triangulation Recognition Heuristics

MCS

LexBFS

MCSM

LexM

These algorithms are guaranteed to compute perfect elimination orderings for chordal graphs. MCSM and LexM are variants of MCS and LexBFS that compute minimal orderings. The Lex algorithms were pubished first, and the MCS algorithms were introducd later as simplications. In practice, these algorithms work poorly on non-chordal graphs.

Bandwidth and Envelope Reduction Heuristics

RCMMD

RCMGL

Spectral

These algorithms seek to minimize the bandwidth and profile of the permuted graph, quantities that upper bound the width and fill of theinduced chordal completion. RCMMD and RCMGL are two variants of the reverse Cuthill-McKee algorithm, a type of breadth-first search. They differ in in their choice of starting vertex. In practice, these algorithms work better than the triangulation recognition heuristics and worse than the greedy heuristics.

Greedy Heuristics

MMD

MF

AMD

SymAMD

AMF

These algorithms simulate the elimination process, greedity selecting vertices to eliminate. MMD selects a vertex of minimum degree, and MF selects a vertex that induces the least fill. Updating the degree or fill of every vertex after elimination is costly; the algorithms AMD, SymAMD, and AMF are relaxations that work by approximating these values. The AMD algorithm is the state-of-the-practice for sparse matrix ordering.

Exact Treewidth Algorithms

BT

These algorithm minimizes the treewidth of the completed graph. Beware! This is an NP-hard problem. I recommend wrapping exact treewidth algorithms with preprocessors like RuleReduction or ComponentReduction.

Meta Algorithms

MinimalChordal

CompositeRotations

RuleReduction

ComponentReduction

These algorithms are parametrized by another algorithm and work by transforming its input or output.

CliqueTrees.PermutationOrAlgorithm — Type

PermutationOrAlgorithm = Union{AbstractVector, EliminationAlgorithm}

Either a permutation or an algorithm.

CliqueTrees.DEFAULT_ELIMINATION_ALGORITHM — Constant

DEFAULT_ELIMINATION_ALGORITHM = MMD()

The default algorithm.

CliqueTrees.BFS — Type

BFS <: EliminationAlgorithm

BFS()

The breadth-first search algorithm.

CliqueTrees.MCS — Type

MCS <: EliminationAlgorithm

MCS()

The maximum cardinality search algorithm.

Reference

Tarjan, Robert E., and Mihalis Yannakakis. "Simple linear-time algorithms to test chordality of graphs, test acyclicity of hypergraphs, and selectively reduce acyclic hypergraphs." SIAM Journal on Computing 13.3 (1984): 566-579.

CliqueTrees.LexBFS — Type

LexBFS <: EliminationAlgorithm

LexBFS()

The lexicographic breadth-first-search algorithm.

Reference

Rose, Donald J., R. Endre Tarjan, and George S. Lueker. "Algorithmic aspects of vertex elimination on graphs." SIAM Journal on Computing 5.2 (1976): 266-283.

CliqueTrees.RCMMD — Type

RCMMD{A} <: EliminationAlgorithm

RCMMD(alg::Algorithm)

RCMMD()

The reverse Cuthill-McKee algorithm. An initial vertex is selected using the minimum degree heuristic.

Parameters

alg: sorting algorithm

Reference

Cuthill, Elizabeth, and James McKee. "Reducing the bandwidth of sparse symmetric matrices." Proceedings of the 1969 24th National Conference. 1969.

CliqueTrees.RCMGL — Type

RCMGL{A} <: EliminationAlgorithm

RCMGL(alg::Algorithm)

RCMGL()

The reverse Cuthill-McKee algorithm. An initial vertex is selected using George and Liu's variant of the GPS algorithm.

Parameters

alg: sorting algorithm

Reference

George, Alan, and Joseph WH Liu. "An implementation of a pseudoperipheral node finder." ACM Transactions on Mathematical Software (TOMS) 5.3 (1979): 284-295.

CliqueTrees.RCM — Type

RCM = RCMGL

The default variant of the reverse Cuthill-Mckee algorithm.

CliqueTrees.LexM — Type

LexM <: EliminationAlgorithm

LexM()

A minimal variant of the lexicographic breadth-first-search algorithm.

Reference

Rose, Donald J., R. Endre Tarjan, and George S. Lueker. "Algorithmic aspects of vertex elimination on graphs." SIAM Journal on Computing 5.2 (1976): 266-283.

CliqueTrees.MCSM — Type

MCSM <: EliminationAlgorithm

MCSM()

A minimal variant of the maximal cardinality search algorithm.

Reference

Berry, Anne, et al. "Maximum cardinality search for computing minimal triangulations of graphs." Algorithmica 39 (2004): 287-298.

CliqueTrees.AMD — Type

AMD <: EliminationAlgorithm

AMD(; dense=10.0, aggressive=1.0)

The approximate minimum degree algorithm.

Parameters

dense: dense row parameter

aggressive: aggressive absorption

Reference

Amestoy, Patrick R., Timothy A. Davis, and Iain S. Duff. "An approximate minimum degree ordering algorithm." SIAM Journal on Matrix Analysis and Applications 17.4 (1996): 886-905.

CliqueTrees.SymAMD — Type

SymAMD <: EliminationAlgorithm

SymAMD(; dense_row=10.0, dense_col=10.0, aggressive=1.0)

The column approximate minimum degree algorithm.

Parameters

dense_row: dense row parameter

dense_column: dense column parameter

aggressive: aggressive absorption

Reference

Davis, Timothy A., et al. "A column approximate minimum degree ordering algorithm." ACM Transactions on Mathematical Software (TOMS) 30.3 (2004): 353-376.

CliqueTrees.AMF — Type

AMF <: EliminationAlgorithm

AMF(; speed=1)

The approximate minimum fill algorithm.

Parameters

speed: fill approximation strategy (1, 2, or, 3)

Reference

Rothberg, Edward, and Stanley C. Eisenstat. "Node selection strategies for bottom-up sparse matrix ordering." SIAM Journal on Matrix Analysis and Applications 19.3 (1998): 682-695.

CliqueTrees.MF — Type

MF <: EliminationAlgorithm

MF()

The greedy minimum fill algorithm.

Reference

Tinney, William F., and John W. Walker. "Direct solutions of sparse network equations by optimally ordered triangular factorization." Proceedings of the IEEE 55.11 (1967): 1801-1809.

CliqueTrees.MMD — Type

MMD <: EliminationAlgorithm

MMD(; delta=0)

The multiple minimum degree algorithm.

Parameters

delta: tolerance for multiple elimination

Reference

Liu, Joseph WH. "Modification of the minimum-degree algorithm by multiple elimination." ACM Transactions on Mathematical Software (TOMS) 11.2 (1985): 141-153.

CliqueTrees.METIS — Type

METIS <: EliminationAlgorithm

METIS(; ctype=-1, rtype=-1, nseps=-1, niter=-1, seed=-1,
        compress=-1, ccorder=-1, pfactor=-1, ufactor=-1)

The multilevel nested dissection algorithm implemented in METIS.

Parameters

ctype: matching scheme to be used during coarsening

rtype: algorithm used for refinement

nseps: number of different separators computed at each level of nested dissection

niter: number of iterations for refinement algorithm at each stage of the uncoarsening process

seed: random seed

compress: whether to combine vertices with identical adjacency lists

ccorder: whether to order connected components separately

pfactor: minimum degree of vertices that will be ordered last

ufactor: maximum allowed load imbalance partitions

Reference

Karypis, George, and Vipin Kumar. "A fast and high quality multilevel scheme for partitioning irregular graphs." SIAM Journal on Scientific Computing 20.1 (1998): 359-392.

CliqueTrees.Spectral — Type

Spectral <: EliminationAlgorithm

Spectral(; tol=0.0)

The spectral ordering algorithm only works on connected graphs. In order to use it, import the package Laplacians.

Parameters

tol: tolerance for convergence

Reference

Barnard, Stephen T., Alex Pothen, and Horst D. Simon. "A spectral algorithm for envelope reduction of sparse matrices." Proceedings of the 1993 ACM/IEEE Conference on Supercomputing. 1993.

CliqueTrees.BT — Type

BT <: EliminationAlgorithm

BT()

The Bouchitte-Todinca algorithm.

Reference

Korhonen, Tuukka, Jeremias Berg, and Matti Järvisalo. "Solving Graph Problems via Potential Maximal Cliques: An Experimental Evaluation of the Bouchitté-Todinca Algorithm." Journal of Experimental Algorithmics (JEA) 24 (2019): 1-19.

CliqueTrees.MinimalChordal — Type

MinimalChordal{A} <: EliminationAlgorithm

MinimalChordal(alg::PermutationOrAlgorithm)

MinimalChordal()

Evaluate an elimination algorithm, and them improve its output using the MinimalChordal algorithm. The result is guaranteed to be minimal.

Parameters

alg: elimination algorithm

Reference

Blair, Jean RS, Pinar Heggernes, and Jan Arne Telle. "A practical algorithm for making filled graphs minimal." Theoretical Computer Science 250.1-2 (2001): 125-141.

CliqueTrees.CompositeRotations — Type

CompositeRotations{C, A} <: EliminationAlgorithm

CompositeRotations(clique::AbstractVector, alg::EliminationAlgorithm)

CompositeRotations(clique::AbstractVector)

Evaluate an eliminaton algorithm, ensuring that the given clique is at the end of the ordering.

Parameters

clique: clique

alg: elimination algorithm

Reference

Liu, Joseph WH. "Equivalent sparse matrix reordering by elimination tree rotations." Siam Journal on Scientific and Statistical Computing 9.3 (1988): 424-444.

CliqueTrees.RuleReduction — Type

RuleReduction{A} <: EliminationAlgorithm

RuleReduction(alg::PermutationOrAlgororithm)

RuleReduction()

Preprocess a graph using safe reduction rules.

Parameters

alg: elimination algorithm

Reference

Bodlaender, Hans L., Arie MCA Koster, and Frank van den Eijkhof. "Preprocessing rules for triangulation of probabilistic networks." Computational Intelligence 21.3 (2005): 286-305.

CliqueTrees.ComponentReduction — Type

ComponentReduction{A} <: EliminationAlgorithm

ComponentReduction(alg::PermutationOrAlgorithm)

ComponentReduction()

Apply an elimination algorithm to each connected component of a graph.

Parameters

alg: elimination algorithm

CliqueTrees.permutation — Function

permutation(graph;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

Construct a fill-reducing permutation of the vertices of a simple graph.

julia> using CliqueTrees

julia> graph = [
           0 1 0 0 0 0 0 0
           1 0 1 0 0 1 0 0
           0 1 0 1 0 1 1 1
           0 0 1 0 0 0 0 0
           0 0 0 0 0 1 1 0
           0 1 1 0 1 0 0 0
           0 0 1 0 1 0 0 1
           0 0 1 0 0 0 1 0
       ];

julia> order, index = permutation(graph);

julia> order
8-element Vector{Int64}:
 4
 1
 2
 8
 5
 3
 6
 7

julia> index == invperm(order)
true

CliqueTrees.mcs — Function

mcs(graph[, clique])

Perform a maximum cardinality search, optionally specifying a clique to be ordered last. Returns the inverse permutation.

Supernodes

CliqueTrees.SupernodeType — Type

SupernodeType

A type of supernode partition. The options are

type	name

Nodal	nodal supernode partition

Maximal	maximal supernode partition

Fundamental	fundamental supernode partition

CliqueTrees.DEFAULT_SUPERNODE_TYPE — Constant

DEFAULT_SUPERNODE_TYPE = Maximal()

The default supernode partition.

CliqueTrees.Nodal — Type

Nodal <: SupernodeType

A nodal supernode partition.

CliqueTrees.Maximal — Type

Maximal <: SupernodeType

A maximal supernode partition.

CliqueTrees.Fundamental — Type

Fundamental <: SupernodeType

A fundamental supernode partition.

Linked Lists

CliqueTrees.SinglyLinkedList — Type

SinglyLinkedList{I, Head, Next} <: AbstractLinkedList{I}

SinglyLinkedList{I}(n::Integer)

A singly linked list of distinct natural numbers. This type supports the iteration interface.

julia> using CliqueTrees

julia> list = SinglyLinkedList{Int}(10)
SinglyLinkedList{Int64, Array{Int64, 0}, Vector{Int64}}:

julia> pushfirst!(list, 4, 5, 6, 7, 8, 9)
SinglyLinkedList{Int64, Array{Int64, 0}, Vector{Int64}}:
 4
 5
 6
 7
 8
 ⋮

julia> collect(list)
6-element Vector{Int64}:
 4
 5
 6
 7
 8
 9

Trees

CliqueTrees.AbstractTree — Type

AbstractTree{V} = Union{Tree{V}, SupernodeTree{V}, CliqueTree{V}}

A rooted forest. This type implements the indexed tree interface.

CliqueTrees.rootindices — Function

rootindices(tree::AbstractTree)

Get the roots of a rooted forest.

CliqueTrees.firstchildindex — Function

firstchildindex(tree::AbstractTree, i::Integer)

Get the first child of node i. Returns nothing if i is a leaf.

CliqueTrees.ancestorindices — Function

ancestorindices(tree::AbstractTree, i::Integer)

Get the proper ancestors of node i.

Trees

CliqueTrees.Tree — Type

Tree{V} <: AbstractUnitRange{V}

Tree(tree::AbstractTree)

Tree{V}(tree::AbstractTree) where V

A rooted forest with vertices of type V. This type implements the indexed tree interface.

CliqueTrees.eliminationtree — Function

eliminationtree(graph;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

Construct a tree-depth decomposition of a simple graph.

julia> using CliqueTrees

julia> graph = [
           0 1 0 0 0 0 0 0
           1 0 1 0 0 1 0 0
           0 1 0 1 0 1 1 1
           0 0 1 0 0 0 0 0
           0 0 0 0 0 1 1 0
           0 1 1 0 1 0 0 0
           0 0 1 0 1 0 0 1
           0 0 1 0 0 0 1 0
       ];

julia> label, tree = eliminationtree(graph);

julia> tree
8-element Tree{Int64}:
 8
 └─ 7
    ├─ 5
    └─ 6
       ├─ 1
       ├─ 3
       │  └─ 2
       └─ 4

Supernode Trees

CliqueTrees.SupernodeTree — Type

SupernodeTree{V} <: AbstractVector{UnitRange{V}}

A supernodal elimination tree with vertices of type V. This type implements the indexed tree interface.

CliqueTrees.supernodetree — Function

supernodetree(graph;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM,
    snd::SupernodeType=DEFAULT_SUPERNODE_TYPE)

Construct a supernodal elimination tree.

Clique Trees

CliqueTrees.Clique — Type

Clique{V, E} <: AbstractVector{V}

A clique of a clique tree.

CliqueTrees.CliqueTree — Type

CliqueTree{V, E} <: AbstractVector{Clique{V, E}}

A clique tree with vertices of type V and edges of type E. This type implements the indexed tree interface.

CliqueTrees.cliquetree — Function

cliquetree(graph;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM,
    snd::SupernodeType=DEFAULT_SUPERNODE_TYPE)

Construct a tree decomposition of a simple graph. The vertices of the graph are first ordered by a fill-reducing permutation computed by the algorithm alg. The size of the resulting decomposition is determined by the supernode partition snd.

julia> using CliqueTrees

julia> graph = [
           0 1 0 0 0 0 0 0
           1 0 1 0 0 1 0 0
           0 1 0 1 0 1 1 1
           0 0 1 0 0 0 0 0
           0 0 0 0 0 1 1 0
           0 1 1 0 1 0 0 0
           0 0 1 0 1 0 0 1
           0 0 1 0 0 0 1 0
       ];

julia> label, tree = cliquetree(graph);

julia> tree
6-element CliqueTree{Int64, Int64}:
 [6, 7, 8]
 └─ [5, 7, 8]
    ├─ [1, 5]
    ├─ [3, 5, 7]
    │  └─ [2, 3]
    └─ [4, 5, 8]

CliqueTrees.treewidth — Function

treewidth(tree::CliqueTree)

Compute the width of a clique tree.

treewidth(graph;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

Compute an upper bound to the tree width of a simple graph.

CliqueTrees.separator — Function

separator(clique::Clique)

Get the separator of a clique.

separator(tree::CliqueTree, i::Integer)

Get the separator at node i.

CliqueTrees.residual — Function

residual(clique::Clique)

Get the residual of a clique.

residual(tree::CliqueTree, i::Integer)

Get the residual at node i.

Filled Graphs

CliqueTrees.FilledGraph — Type

FilledGraph{V, E} <: AbstractGraph{V}

A filled graph.

CliqueTrees.ischordal — Function

ischordal(graph)

Determine whether a simple graph is chordal.

CliqueTrees.isperfect — Function

isperfect(graph, order::AbstractVector[, index::AbstractVector])

Determine whether an fill-reducing permutation is perfect.